Make Hurl Torchlight not damage the player, increase range 4->5
The player was already immune if they were undead or non-living, but while
I'd hoped in general that not being able to use it at point-blank range
would make positioning for it more interesting, in practice I'm not sure
this is happening (or that many zombie pileups allow *any* positioning that
could do something useful without hitting the player). So for now, let's
try it just being harmless to you.