Don't look at movement delay when computing Vhi's Electric Charge delay
Spriggans, nagas, and Chei characters don't take an unusual amount of time
for any other translocations. Vhi's is more controlled than Blink or
Passage (and Passage does usually involve some walking to actually use),
but it is still restricted enough in use that I don't see a problem with
letting it be a fast way for nagas or Chei characters to reposition in
certain situations. If necessary, it seems much easier to justify Chei
hating this spell than to justify this spell looking at movement delay.

There were also some weird interactions previously with Vhi's being slower
to cast for humans in shallow water and faster with swiftness active, etc.