(Hopefully) fix some crashes when reviving apostles that died in portals
The game was complaining that an Ice Cave's exit didn't generate, which may
have been because it was already used to leave from (and closed).

At any rate, since the player should normally never be able to go back to
the same floor of any disconnected branch again, it should be fine not to
attempt to do cleanup on them... right?