Don't crash on saving game after abandoning Beogh with apostles
The dump was trying to get the name of the 'Dismiss Apostle' abilities so
that it could show actions counts in the dump and with all apostles now
erased from their internal slots, was crashing.

Also: make logs count the 3 Dismiss Apostle abilities as the same ability,
for action count purposes. This code is kinda ugly....