First fresh vault review for 0.32
The usual scaling sets much more versus depth, reducing long tails of
absurd rolls for monsters, cleaning up headers to fit a more consistent
house style, so on and so forth. One particular point that'll fit into
these recurring vault reviews from hereon is that with more decorative
features added and to be added, early crystal appearances and granite
statue counts will be regularly reduced to compensate- there's far too many
uses of such in early D for them to stand out as decorative any more or for
it to stand out in dangerous vaults past the earlygame, which can do with
some trimming down in exchange for some of these fancier options.

A few other things of note:

 * Some hells ambushes have been nerfed a little, with added cover and
   slightly lower monster counts. While extended can and should be quite
   cruel to players, there's a big difference between even the recent
   branch buffs and immediately ambushing a player with adjacent paralysis
   and damnation or making melee characters wade through 5+ curse skulls
   in an open layout, considering the

 * evilmike_ambush is now nerfed in a variety of branches, by using
   hand-picked spawns usually from the given branch it spawns in. The sheer
   hostility of the vault was less of a concern when far more of the later
   branches had far more harmless spawns in them, but this was heavily
   aggravated by the revision to hell's sets and the introduction of
   things like Blink Allies Encircling and berserkable tmons. Most of the
   branch placements still have sheer volume and a few notable threats
   at such high quantities, anyway.

 * Air zigs now use wind drakes and ironbound thunderhulks, to add a little
   more variety.