Let orc apostles walk on water
Or something close enough to it, anyway.

They basically count as amphibious in all ways, aside from being drawn
on top of water, not resisting drowning, and having different xv lines.
The code for this feels slightly hacky, but is minimal and localized, so
probably fine.

Swamp and Shouls have continued to be awkward enough for Beogh worshippers
that I think it's mechanically important to do this (and the flavor fits!)