New decoration: mystic cages
In the footsteps of conduits and on a different track than the divine and
demonic statues prior, these are meant to start providing inobtrusive
decorative features reflecting a relatively more regular life inside of
the dungeon beyond mostly empty rooms filled with enemies milling about.

Mystic cages are metal statue redefinitions as kennels and enclosures for
many pets: the hounds of gnolls and orcs, the menageries of vampires, the
experiments of various wizards, the hypothetically tamed beasts within
Vaults, and so on and so forth. With how often there's a mix of given
intelligences in randomized generic floor spawn sets, it's easy to miss
a vault deliberately pairing given monsters together if noise makes them
wander or split- these can help straightforwardly imply such with minimal
effort.

The mystic cage tiles are a mild edit of the tile for Maxwell's Etheric
Cage, made by CanOfWorms, removed back in 0.25.