Split monster Vhi's into Vhi's Electrolunge
Functionally very similar to player Vhi's except that it doesn't care about
the distance travelled for its damage, and simply applies a flat (HD-based)
amount of electric damage if the initial attack hits (rather than being a
multiplier on the weapon damage).

It has a different name in large part to give it a different description
when examined. xv properly shows the added damage.

(I also gave it +1 range compared to player Vhi's)