Remove Yred piety bonus for mostly-cleared floors
This bonus was extremely small in practice (only +~1.5 extra piety for
clearing 91% instead of 66%) but multiple players assumed it was larger and
warped some of their play decisions around it. This isn't a desired effect,
so I'm simply removing the bonus entirely (but leaving the flavor messages)

If this somehow proves on overly meaningful nerf to Yred's overall piety
gain, we can tweak it up in another way.