Leave allies behind when taking transporters without room to place them
The current behaviour when taking a transporter is that the game tries to
place all monsters coming with you near the player, and then if there is
not enough room for them, to place them anywhere else in the level. In some
rare situations this has awkward effects: for instance, if there isn't
room to place followers after a Gauntlet transporter, those followers can
be placed in inaccessible bubbles of the Gauntlet instead. This is
especially bad for Beogh worshippers, as it can lead to their apostles
immediately dying in ways you will never be able to avenge.

Instead, if there is no room to place a friendly follower near you, simply
leave them behind.

Note: a similar thing happens when travelling across stairs, and while it
also makes sense not to have elevator vaults teleport allies across a
floor, I don't think it is safe to do so in this case; we cannot choose to
leave a creature behind at its original location because that floor is
already unloaded - they would simply cease to exist entirely.