Rewrite backend handling of the player's apostles and their state
The original version of this used a whole lot of different props (and even
assembled some prop keys via a prefix and the monster's mid) and was
generally a little unwieldy (even if it did work). I have converted this
over to using a small data structure specifically for tracking apostle
state. This generally simplifies code and will probably provide future
benefits.

It also allows me to track a 'vengeance contribution' for each individual
apostle the player has, which I am using to adjust the revival formula yet
again to fix a few weird cases (eg: dismissed apostles would still
contribute a vengeance bonus upon vegeance completion, and the way piety
needed was capped could cause 3 deaths to be cheaper than 2).

**THIS WILL BREAK SAVES WORSHIPPING BEOGH PRIOR TO THIS POINT**

Or rather, players with Beogh exprimental saves that upgrade to this
version will probably lose the ability to revive apostles they have, and
other things may go wrong. But this way, things on the backend should be
better by the time we merge with trunk.

I have tested most of the interactions affected by this change, but have
no confidence I have squashed 100% of the bugs. It was tricky to set up in
the first place, so we'll have to see if people find more of them....