Give Recall Apostles at *1 instead of *3, adjust keybinds
The player cannot *get* apostles until 3*, but it felt awkward for
smiting to sometimes remove the ability of the player to call existing
apostles between levels. Now the player gets it at 1* (but the ability
will be hidden if they have not recruited any yet)

Also adjust internal ordering so the default keybind for recall and smite
aren't so weird.