Allow vampires to always shapeshift
They are known to turn into various creatures and mist in folklore, and
are also able to mutate normally already, so it wouldn't be weird if
their bodies were malleable enough to be polymorphed or shapeshift
while bloodless.
Also, it'd make them more fun and unique as a shapeshifting undead
species.