Slightly refactor chaos_affect_defender to fix a couple of bugs
Chaos_affect_defender was a member of attack, and was used for
chaos-branded weapons, but it was ALSO used for chaos clouds. And this
meant that clouds had to construct a fake attack just to call it. This had
the consequence that there always had to be an 'attacker'. Since clouds
often have no agent, the game defaulted to having the 'attacker' be the
actor standing in the cloud.

This was capable of causing penance in the following way. A friendly
monster steps into the cloud and gets confused by it. Since they are the
'attacker' for this chaos effect, they count as having been confused by a
player-friendly source for purposes of blame tracking. (This is how the
player gets credited for kills done by monsters they confused). But then
when this confused monster triggers the cloud again, because the source of
the chaos effect (ie: themselves) was confused 'by you', YOU are considered
the source of the new debuff they recieve, which counts as you attacking
them yourself, as far as the good gods and Beogh are concerned.

I have pulled chaos_affects_defender out attacks entirely. Now it is
chaos_affects_actor() and the source is optional. It should otherwise
operate basically the same as it did prevously, aside from fixing an
additional bug where web effects from a chaos weapon incorrectly re-applied
the damage that attack did and produced weird messages.

(Note: I wasn't sure what file to put this in, so it's ended up in
clouds.cc. If someone feels strongly, they could put it somewhere else, I
guess)