Burning bush -> scrub nettle (monkooky)
There's been some concerns with the increase of decorations that some
non-decorative monsters shouldn't be confusable with decorative features.
Burning bushes, those weird biblical jokes, fall directly under this
concern. A regular bush can already be set on fire and use the same tile as
the other enemy in question without suddenly picking up the ability to
shoot 3d9+ Throw Flame. They're only used in one sprint, one vault's
configuration for one branch, and by Fedhas's Overgrow ability, so
modifying them into something a bit less confusable should address this
without upsetting balance too much.

The new scrub nettle is an Australian nettle, mildly enhanced in the way
dart slugs and bombardier beetles are. It uses Sting over the burning bush
Throw Flame, which gives them a partially irresistable attack that actually
scales notably directly with one's Invocations skill (since orc apostles
can get the spell for a while anyway). In exchange, the resistance is
vastly more common, they're lower HD / HP to start with, and Sting only has
a range of 4 instead of 7 (not that this is any more awkward as an ally
than the wandering mushrooms the ability already grants). This probably
doesn't upset Fedhas balance too much- Overgrow is very expensive and used
for terrain-manipulation as much as it is for turret and mushroom support.
We could look into invocations not granting such a high ratio of invo
skill versus HD _every_ plant Fedhas uses, if needed.

It might be useful in the right heavily-constrained portal, as a
introductory baby immobile in a context of Sewers already occasionally
having strange numbers of plants. If the name is somewhat questioned for
not seeming threatening while the bulk of immotile plants aren't, it can
use the far more direct name of "stinging nettle" instead.

Their placeholder tile is a deeply silly mix of the original rltiles spiny
worm, the first tile for porcupines by Malwyn, and an unused "poison
pollen" plant tile from wad left in UNUSED for twelve years.

(If these don't work out, it might be possible to set a different plant on
fire instead of bushes, or maybe make a monster that's the base for a
solitary snaplasher vine to reach out?)