New decorations: sacred lotuses
The opposite of demonic plants in multiple ways, sacred lotuses are holy
water-bound plants that don't wither away and have sturdiness even beyond
bushes. They provide some additional decorative plants of a more notable
nature than the old proliferation of trees and plants themselves, and are
meant to only be used in very sparing pillars and sides rather than
potential blockades like most other non-hostile plants are delegated to.

They're used in vaults for Elyvilon (as symbols of longevity), Fedhas (as
a revered plant), and the Wu Jian Council (as a symbol of elegance and
grace). In particular, the first of those provides a little more fanciness
for holy themed vaults, which focus heavily on wall and statue retiling
due to the rather limited breadth one can use holies when compared to the
wide variety of potential unholy themes. It can also provide a little bit
of fanciness to the very watery, plant-filled branches of Swamp and Shoals.

They use the old thorn lotus tile by snw-0.