Make companion_is_elsewhere lie less
Companion_is_elsewhere relied on a manually updated list of what floor each
companion was supposed to be on. This was normally pretty reliable, but
there were a few cases (the Abyss in particular) where the list could THINK
they were there, but were not (eg: because the map was regenerated since
the player was last there)

Some results of this:
 -Being unable to recall followers that the game THOUGHT were on the floor
  with you, but were not
 -A particularly troublesome bug where old dactions to upgrade a Hep
  ancestor would incorrectly think the ancestor was on the current floor
  (since it wasn't 'elsewhere'!) and then try to change its weapon.
  Except, since it was referring to a monster that wasn't REALLY on the
  floor, trying to access its weapon would point to a random (probably
  undefined) item on the floor's item list instead. Tinkering with this
  could create various forms of buggy item to appear and crash the game.

Now, since we know we can't 100% trust the cached list, whenever the
list thinks a companion is on our current floor, we double-check that it
actually is. And if it is not, set its location to an invalid floor (since
it's currently effectively 'nowhere')

This should fix both above bugs and possibly some others.