Adjust Zig set scaling a little more
Emphasize spriggan druids in Lair floors a little less, scale Tomb / Pan /
Fire floors a little harder, let the non-Tar hells at the higher ends
emphasize a little more of the selected monsters, and screw with some
experimental quadratic scaling for Air (titans) and Earth (hell sentinels
and iron giants). Further steps still need to be taken, but this continues
to be a very quick way to continue to adjust megazigging.