Fix dactions that could cause Ancestors/Bound Souls to incorrectly die
Because entering a new floor causes all dactions from the entire game's
history to fire (see 063869c3723e ), the dactions for losing your Hep
ancestor (due to piety loss/excommunication) or bound soul (due to binding
a new one) could happen repeatedly. This usually did nothing, but in the
Abyss in particular could cause them to simply die - a minor nuissance for
an ancestor, but a permanent mysterious loss for a bound soul.

Now we try to be a bit more picky about whether we actually kill something
with this daction: check that the player still doesn't qualify for an
ancestor, or that the bound soul we're trying to remove isn't the player's
current companion.