Fix bound souls being saved to the companion list as hostile
This was seemingly always the case, but usually not noticable as each
level transit would update the saved copy (with the real, friendly one).
But in some cases where a soul didn't transit properly and was recalled,
it would be recalled as hostile and remain hostile forever afterward -
even after being repeatedly killed!

(It was theoretically even possible to farm unlimited items off a bound
soul this way, by recalling and killing it over and over again)