Don't transit monsters marked as followers
Normally when entering a floor, followers (ie: things taking the stairs
with you) are placed before transiting monsters (ie: things that fell down
a shaft earlier or were banished). But Abyss generation could place
transiting monsters (sometimes multiple times?) before the normal call to
place followers.

Followers are placed near the player, while transiting monsters are placed
anywhere, so this could result in player allies seeming to disappear when
you stepped through an Abyss portal.

This also had the unfortunate side-effect of not updating the companion
list properly, so let's try to prevent it.
but transiting monsters placed randomly on the floor.