Unbreak telling allies to guard an area
This was broken by 0a82c8690a01462b55615d2cbd8e7497dc3b9d6c changing the
LoS check for allies to return to you from 'Does this ally currently have
LoS to the player' to 'Does this ally ever need LoS to the player to
attack', resulting in all perma-allies attempting to return to the player
every single turn - whether they could see you or not.

I've also carved out an additional exception to not return to you if you
told them to guard an area, since this is probably what the player wants
in that case.

Additionally, telling allies to follow you, or recalling them, should fully
end their travel behaviour. (Previously - and unrelated to the linked
commit - telling them to guard an area after you'd previously told them to
guard a different area would often result in them trying to run back to
the first one or something outright being stuck in place, since they would
still be trying to use pathfinding info from the previous time they did
pathfinding)

I am 100% positive travel_path is still not being cleared in all situations
that it should be, so it's possible that the minor change in
_handle_monster_travelling will uncover some additional bugs, but the code
currently seems to assume that the monster's target will be travel_path[0],
if it's travelling, and thus if anything sets its target directly, it can
cause it to perpetually freeze in place by assuming it's moving to its next
waypoint when it's not actually moving anywhere at all. Probably better to
find any other bugs as they come up rather than leave this one here.

(Probably better to refactor a lot of monster movement / behaviour code,
but HOO BOY is that a tall ask...)