New blast wand: Wand of Warping
Alternates with Iceblast and Roots.

This implements one of the discussed proposals for an alternate AoE attack
wand to make /roots a little less common. It does a similar fireball-sized
blast and has a chance to blinks everything caught in the blast radius
(including you!)

It does slightly less damage than iceblast (though is irresistable), but
while blinking enemies has strong utility potential, in practice the
scattering of them can be as much of a drawback as it is an advantage.
Sometimes you can use it to juggle something away from you, other times it
will bring them closer or reduce your ability to hit multiple targets.
Being able to deliberately hit yourself with it as a way to escape some
situations (at the cost of damage/charges) also feels like a reasonable
part of this package. Sometimes more unwieldy than other wands, sometimes
more useful, and definitely distinct.

The blinking chance starts at 50% and scales to 90% by ~15 evocations.

Make /warping not 100%