Save the Djinn!
Instead of complete random spells from all sets that make the Djinn
nearly completely unplayable, give the Djinn random spells from a
reasonable set, where this is a suggestion.

Consider characters based on Fire, Ice, Air, Earth and Alchemy as
elementalists and they get random skills from within those sets, as you
need more than one element to defeat various enemies that are immune to
attacks from one or more elements.

Conjurers get their skills from the Conjurer set alone that is quite
large.

Hexslingers get spells from the Hexes and Ice schools, with added spells
Infestation and Shatter to patch the set.

Summoners gets spells from Summoning set and Earth set as otherwise the
character will not be viable, and the Earth set is a powerful
supplement. Moreover, Summon Ice Beast is used for a level 2 spell as
there is no additional level 2 spells besides the initial spells from
the Summoning spell set.

Necromancers are based on the school of Necromancy spell set only, but
since there are several spell levels that might not have an option, it
is heavily patched. The patched spells used are:
Mephitic Cloud, Iskanderun's Mystic Blast, Orb of Destruction and Polar
Vortex.

Warpers use spells from the school of Translocation, but as with the
Necromancer there is a need for many patched spells. The patched spells
used are:
Irradiate, Spellforged Servitor, Maxwell's Capacitive Coupling, Chain
Lightning and Dragon's Call.

Enchanters were difficult to define, and as they are not elementalists,
the decision was to use spells from the Hexes, Necromancy, Translocation
and Conjuration spell schools, thus giving them a wide range of spells.

The characters have been tested (not to the highest skill) and are fine
playable, with the Warper being the toughest as the Warper is partly
melee, but the Djinn is not the best melee species.

Other character setups are unaffected.