Change how the Repel Missiles effect works
This is both a simplification and a rebalancing.

Current, repel missiles works by reducing the to-hit roll of a missile by
a random amount from 0-100% of its current value (or 0-50% of its value if
it's a penetrating beam). In effect, this acts a little like a multiplier
of the defender's EV - it does almost nothing at low EV and its
effectiveness directly rises with EV - but with different math.

This has several arguably bad effects:
 -The buff has little effect on enemies with mediocre EV (and there's been
  a desire to make an rMsl aura enemy, who would currently accomplish
  little when supporting many types of enemy)
 -The effect of the buff is *enormously* counteracted by EV reductions like
  wand of roots (which used to grant something like a +2800% chance to hit
  versus The Enchantress, as an extreme example). If enemies have rMsl, I
  think you should *feel* it in the same way that you feel enemy rF when
  you're attacking with fire
 -Scarves of repulsion were close to useless on many characters, but a lot
  of players did not realize this and assumed they were still providing
  meaningful bonuses to characters with no EV. This new system should make
  that a lot clearer to explain, as well as make some itemization choices
  viable that previously were not.

This changes rMsl to be a flat +EV versus missile attacks instead, and
operates even if the target is incapacitated (the magic is still pushing
arrows away, even if the caster is asleep!) The old behaviour of being
only half as efffect against penetrating beams is removed also (SH already
offers a means of attack avoidance that specifically is worse against
penetrating beams).

In enemy hands, it should have a stronger effect on most enemies, and be
much less negated by constriction (though constriction still helps a bunch
in most cases!). In player hands, this is a buff if the player is below
~18 EV and a nerf if the player is above that (which I think should make
scarfs of repulsion more appealing to more characters and maybe make
itemization decisions a little more diverse)

Numbers can, of course, be adjusted (and the player/enemy bonus separated),
but I think it's fair that rMsl feel impactful against the type of thing it
resists in a similar way that a pip of resistance would be against its.