Increase Foxfire Swarm damage
Moving it from 1d6 to 1d8 might make it a bit more competitive with
Ignis's other active. Elliptic suggested reducing the max number of
foxfires from 24 to 8 (ie going from radius 2 to 1), which *would*
make it easier to balance, but would be a terrible loss to the
spectacle of the effect. I'd rather hold off.