feat: Specifying fixed properties for randarts
This commit adds support to item_spec to specify an arbitrary set of
artefact property values. The values are given in any item spec
according to the following syntax:

artprops:prop1:val1&prop2:val2 ...

Matches against property names are case insensitive, and properties like
rPois and ^Contam that are boolean can omit specifying the value. This
syntax can be used in any item spec, including when defining items for
vaults, giving items to monsters, etc.

When making the item, the randart property randomization considers the
fixed values in terms never overriding these values. Randart generation
considers how many good/bad properties the fixed ones add to the item
and proceeds with its usual randomization process, which means that it
may add additional random properties.

Some error checking is done on the given value ranges to prevent
negative values on properties that are only ever positive and to only
allow 0 or 1 on boolean values.