Allow previewing EV change for armour in item descriptions
This is a much more complicated question than displaying AC changes was,
since EV depends on a complicated formula that hinges not only on the
change in ER between the two armour being compared, but also differences in
str/dex from properties on the two items.

Rather than try to duplicate the complex logic of _player_evasion() or
parameterize every single aspect of it (which would be brittle and
error-prone if any aspect of gear properties ever changed), I have opted
to 'simulate' the correct answer by internally equipping the item just long
enough to calculate player EV while wearing it.

As this is done by directly manipulating equipment links, it should be
invisible and have no side-effects rewarding wielding/unwielding.

One complication is that Crawl's item system makes it literally impossible
for the player to equip (even internally) an item that isn't in their
inventory. And since inventory is capped at a fixed size and might be full
(and the player might be examining an item for a slot which is currently
empty!), there is no normal way to gaurantee we have anywhere to put this
item, even temporarily.

So I have 'cheated' by adding an extra slot to you.inv beyond MAXPACK (the
normal cap). This should be invisible basically everywhere, and is not
marshalled (since it's only temp working space). This still makes me
slightly nervous, but I have looked through as much code that directly
manipulates this array as I can without seeing a problem this will cause
(other than one fixed by this commit itself), and I also see no other
reasonable way to do this without overhauling item links entirely. Which,
while theoretically nice, is wildly out of scope here.

As this process is still somewhat sketchy, I've tried to document the
gotchas quite explicitly, where they exist. But I think the QoL improvement
to players is well worth the hack.