Interrupt autorest and travel when allies take damage
Because autorest interrupts on fewer things than resting manually does, it
was possible not to be notified of an apostle sitting in a poison cloud
and have it immediately die with one button press. This is obviously quite
bad.

Now interrupt when any visible ally takes damage, to avoid this.

I am... more than a little worried that this will create needless
interruptions in regular play and that this should be more conservative,
but the situation it's aimed at fixing is very bad, and it's been argued
that the downsides won't come up in practice. So let's see, and then maybe
try to narrow down criteria later, if it turns out there's issues.