Don't interrupt rest when allies take poison ticks
This was additionally confusing since there isn't even any message for them
taking damage this way, so it's not clear what's interrupting the player.
(And it happens a lot from incidental damage in places like Snake)

And while it's sad to lose an ally that way, the player also usually can't
do anything about this - if they're lethally poisoned, they're just going
to die. (Standing in a poison cloud does direct damage and will still
interrupt, so most ways to get *more* poisoned will still alert the player)