Remove tentacle damage sharing, remove ink cloud
At 1/4 the damage taken by tentacles, the shared damage to the head was largely irrelevant. For simplicity, remove it.

Ink cloud has long been nonfunctional; unless the player was very stealthy, the Kraken would stay in place and could be blasted from out of LOS. Absent clever ideas to make it work, remove it.