Fix Vhi's destination checking being player-centric (Sergey)
Even if a monster was casting Vhi's, it used only the player to determine
whether a destination tile was safe. This meant that flying monsters would
refuse to dash onto lava if the player was not flying, but also that land
monsters would happily dash into lava if the player *was* flying (and then
crash).

Thank you for your sacrifice, Bogfrid. You will be remembered.