feat: Simplify formulas for artefact properties (elliptic)
When making an item randart, there's a binomial(7, 0.3) distribution
used to assign quality from a level of 1 to 7. This is currently done by
taking a binomial with 7 trials and taking a max of this and 1. This
commit simplifies this to just be a 1 + binomial(6, 0.21), which has a
similar mean.

Additionally this commit simplifies the distribution for the number of
bad artefact properties added. The current implementation is a min of a
binomial distribution and the max allowed bad properties, which is 2
minus the number of any fixed bad properties. This binomial has a number
of trials randomized based on the quality level and random rounding,
with a fixed `p` of 0.3. The new implementation removes the min and the
dependence on quality in the binomial, making the number of trials
simply be the max allowed bad properties. The `p` is reduced to 0.21 to
keep the mean number of bad properties similar to as before.

This distribution will lead to more randarts having only 1 bad property
instead of 2 and will even out the numbers of bad properties across
randarts by quality. It's not clear that the original relationship of
bad props to quality was necessary and this new distribution is much
easier to reason about.

Here are stats for `any randart` from 10k iterations each both before
and after this distribution change:

Before:
Max # of props = 6, mean =  2.87, median = 3
Max # of good props = 5, mean =  2.45, median = 2
Max # of bad props = 2, mean =  0.42, median = 0
Max (good - bad) props = 5, avg # =  2.03

After:
Max # of props = 6, mean =  2.97, median = 3
Max # of good props = 6, mean =  2.54, median = 3
Max # of bad props = 2, mean =  0.44, median = 0
Max (good - bad) props = 6, avg # =  2.10

The means are quite similar for both types of props, so this change
shouldn't have a large effect, but will make future reasoning about
randart balance easier.