Don't apply -Move on Coglin if there are no monsters left
Try this out since this seems responsible for a fair proportion of the
annoyances, e.g. players immediately wanting to manually move at the end of
a fight. This commit also tries to wipe existing -Move that might still be
remaining from a previous Coglin action if the action that concluded the
fight was fast (e.g. a fast melee attack). (In principle maybe this part
should be implemented somewhere else so that -Move is also removed if the
action concluding the fight was a fast action that doesn't normally cause
-Move (e.g. swapping a ring and having an ally kill the last enemy in that
0.5 time), but that seems super rare so this should be sufficient for
testing the change.)

Note that invisible monsters still count for this, so this does leak their
existence - this seemed better than alternatives.