Fix reflected projectiles causing the player penance if they hit an ally
If the player fired a projectile at an enemy who reflected it back, and
the it missed the player and struck an ally behind them, this would be
misattributed to the player themselves (and cause penance for Beogh / good
gods).

Change the conduct check to look at whether the agent() is the player
(which properly handles reflection blame) rather than the original
killer_type.