Make lucid gain more consistent
Use mon props instead of an ench. This should fix being able to gain
lucid from killing a monster with a spell the turn after you melee it
and lucid sometimes not being given from retaliation kills.

Also gain lucid even if your attack was blocked or dodged so if a
monster you're attacking died you don't get confused by not gaining
lucid because it died to something else.