Coglin Gizmos!
While the latest incarnation of the coglin downside (no rings, but
rF+/rC+ at xl 13) may be at a reasonable power level, there's been much
interest expressed in making this system a little more involved,
randomized, and 'exciting'. So here is an attempt at this which tries to
capture some very fun coglin flavour in the process:

Coglins no longer gain fixed resists at xl 13, and in addition to being
unable to wear rings, also cannot wear amulets. But in exchange, they
get a one-time choice of 3 random 'gizmos' at xl 14 - randomized
equipment designed to fill part of the role of the jewelry they replace.

The intent is that these gizmos are still overall a net downside
compared to normal jewelry slots (with the earlier power, flexibility,
and lategame artifact possibilities those hold), but that they preserve
some of the choice and excitement of finding new items and deciding
between them - and not always knowing what you're going to end up using
at the start of the game. A downside that *feels* like an upside. After
all, being fun should be one of the foremost criteria for a new
species, right?

These gizmos are generated as follows:

Each gizmo has exactly 1 'rare property' and 2 'common properties'.

There are 6 total common properties (rF+, rC+, rPois+rCorr, rElec, Will+,
Slay+3) and the player is gauranteed to see all 6 of them each game - just
in randomized combinations.

There are currently 7 different 'rare properties', of which the player
will be offered a random 3 each game - so that you *can't* know in
advance which you will be offered (nor rely on getting a specific one
each game). Several of these draw from existing equipment properties and
several are entirely new. I feel strongly that it helps the player
fun-factor of the system a lot if some of what you get from gizmos
doesn't just feel like "Jewelry, but less of it" but rather contains
some *new* choices instead of old ones (even if they're overall less
powerful).

The list of current rare properties is:

1) Regen + RegenMP

2) Repel Missiles + Clarity

3) Rampage + Acrobat

4) Gadgeteer:
     30% faster evocable recharge rate and 30% chance not to consume a wand
     charge on firing

5) ManaRev:
     RegenMP++ but *only* while fully Revved

6) ParryRev:
     AC+5 while fully Revved, along with a small chance to disarm an enemy
     upon hitting them. (This does not do any damage, unlike Captain's
     Cutlass's extremely significant damage, to be clear)

7) AutoDazzle
     Small chance to fire a non-damaging beam of blinding light at an enemy
     whose attacks you dodge - even at range.

The idea is that different modifiers should appeal to different sorts of
characters, and also that several of them actively lean into the main
gimmick of coglins by synergizing directly with Rev and attacking.

Obviously the balance of all of these is provisional, but I think given
player response to various previous downsides, it's important to try and
sell the *fun* of this foremost. Numbers can be adjusted down as-needed
afterward.

I hope people enjoy them!

(Credit to regret-index for the tilework, and multiple other devs for
design discussion and suggestions)