New decorations: (unmoving) boulders
We should be very sparing on spreading almost any of various monsters used
to make various spells work. Several of them are very weak as threats (fire
vortices, ball lightnings), and several of them are essentially adding new
monsters that have to go over rigorous design criteria entirely seperate of
their functionality for a spell- for example, we're almost certainly not
going to make dryads or cactus giants anything besides glassed monsters in
a gimmick portal if every.

That said, Brom's Barrelling Boulder makes an extremely physical object
that, when made to not move as the default placement works, can reasonably
fit in any number of vaults already placing stones, large rocks, crumbled
columns, shafts, and so on looking for earthen decoration. The boulder has
a different tile when it's not rolling, and uses "dbname:still_boulder" for
a description when not moving. Do note that they should not be used for any
meaningful walls, however, since we've already got normal plants covering
most of that design space already.

The placeholder tile is an extreme edit of canofworms's Brom's spell icon.

By the way, new rule: no sokoban vaults.